Split materializer
