+++ lib/materialist/materializer/internals.rb
@@ -1,0 +1,13 @@
+require_relative './internals/field_mapping'
+require_relative './internals/link_mapping'
+require_relative './internals/link_href_mapping'
+require_relative './internals/class_methods'
+require_relative './internals/dsl'
+require_relative './internals/materializer'
+
+module Materialist
+  module Materializer
+    module Internals
+    end
+  end
+end
