+++ lib/materialist/materializer/internals/field_mapping.rb
@@ -1,0 +1,14 @@
+module Materialist
+  module Materializer
+    module Internals
+      class FieldMapping
+        def initialize(key:, as:)
+          @key = key
+          @as = as
+        end
+
+        attr_reader :key, :as
+      end
+    end
+  end
+end
