+++ lib/materialist/materializer/internals/link_href_mapping.rb
@@ -1,0 +1,14 @@
+module Materialist
+  module Materializer
+    module Internals
+      class LinkHrefMapping
+        def initialize(key:, as:)
+          @key = key
+          @as = as
+        end
+
+        attr_reader :key, :as
+      end
+    end
+  end
+end
