+++ lib/materialist/materializer/internals/link_mapping.rb
@@ -1,0 +1,14 @@
+module Materialist
+  module Materializer
+    module Internals
+      class LinkMapping
+        def initialize(key:)
+          @key = key
+          @mapping = []
+        end
+
+        attr_reader :key, :mapping
+      end
+    end
+  end
+end
